requirements.txt
beautifulsoup4==4.8.0
Pillow==6.1.0
praw==6.3.1
PyYAML==5.1.2
requests==2.22.0
